probe: the calibration read integrates the ACTION code too — `_calibration` had fixed u at argmax q(u|h,a) and integrated chance under it, a read conditional on one code where the deployed search samples both u and z (docs/pre-mcts-validation-2026-09-07.md); now the taken action's `--calibration-actions` (8) most probable codes with renormalised weights (`top_action_codes`, `action_retained_mass` / `action_support` reported) and the chance prior under each — enumerated exactly or sampled — give `expect` = E_u E_z V(g(h,u,z)), the conditional read stays under its own label (`expect_mode`), `sample` is one joint (u,z) draw, `expect_sigma_u` beside `expect_sigma_z`; DECODES carries `expect_mode`; the finite-difference decode test asks for f32 explicitly (x64 is off in the suite); ledger row extended

+++ tests/test_transition_model.py
@@ -486,8 +486,10 @@ def test_exact_decode_loss_is_conditional_entropy_and_matches_finite_differences
     read = exact_decode_loss(distinct, cell_valid)
     assert float(read.loss) < 0.05 and float(read.accuracy) > 0.98
     assert float(read.mutual_information) > np.log(num_valid) - 0.05
+    # f32 on purpose: the suite runs with x64 off, so a float64 request
+    # would silently become f32 anyway.
     random = jnp.asarray(
-        np.random.default_rng(1).normal(size=(6, num_codes)), jnp.float64
+        np.random.default_rng(1).normal(size=(6, num_codes)), jnp.float32
     )
 
     def loss_of(logits):
